+++ src/semantic/types/expression.rs
@@ -58,7 +58,9 @@ impl Typed for Expression<'_> {
                             })),
                         }
                     }
-                    _ => unimplemented!("Get datatype of an expression that resolves into a function call"),
+                    _ => unimplemented!(
+                        "Get datatype of an expression that resolves into a function call"
+                    ),
                 }
             }
             Expression::UnaryOperator(op, exp) => {