+++ src/syntax/parsers/binding.rs
@@ -125,6 +125,26 @@ impl<'a> Parseable<'a> for VariableBinding<'a> {
             }
         };
 
+        // After the expression there should be a new line
+        // to terminate the statement
+        match tokens.get(next_pos) {
+            Some(t) if t.token_type == TokenType::NewLine || t.token_type == TokenType::EOF => {
+                // continue
+            }
+            Some(t) => {
+                return Err(ParsingError::Err(SyntaxError {
+                    error_start: t.position,
+                    error_end: t.get_end_position(),
+                    reason: format!("Unexpected token `{}`, expected a new line", t.value),
+                }))
+            }
+            _ => {
+                // this should never happen, the lexer always appends
+                // an EOF
+                unreachable!("got to the final of a token stream without finding EOF")
+            }
+        } 
+
         let binding = VariableBinding {
             datatype,
             identifier: &identifier,
@@ -348,6 +368,24 @@ mod tests {
                 assert_eq!(16, error.error_end);
                 assert_eq!(
                     "Expected an expression after the equal `=` operator",
+                    error.reason
+                );
+            }
+            _ => panic!("Error expected"),
+        }
+    }
+
+    #[test]
+    fn should_error_when_there_is_no_delimiter() {
+        let tokens = get_tokens(&String::from("val identifier = 322 print(identifier)")).unwrap();
+        let binding = VariableBinding::try_parse(&tokens, 0);
+
+        match binding {
+            Err(ParsingError::Err(error)) => {
+                assert_eq!(21, error.error_start);
+                assert_eq!(26, error.error_end);
+                assert_eq!(
+                    "Unexpected token `print`, expected a new line",
                     error.reason
                 );
             }